+++ pkg/analyzer/CHANGELOG.md
@@ -2,6 +2,9 @@
 * Added the getters `isDartCoreList`, `isDartCoreMap`, `isDartCoreNum`,
   `isDartCoreSet`, and `isDartCoreSymbol` to `DartType`.
 * Added the method `DartObject.toFunctionValue`.
+* Deprecated the `isEquivalentTo(DartType)` method of `DartType`.
+  The operator `==` now correctly considers two types equal if and
+  only if they represent the same type as defined by the spec.
 
 ## 0.37.0
 * Removed deprecated getter `DartType.isUndefined`.